+++ src/maybe/index.ts
@@ -87,35 +87,35 @@ export const fromNullable = <T>(value: T | null | undefined): Maybe<T> => {
   return value == null ? new Nothing<T>() : new Just<T>(value);
 };
 
-export const mapM = <T, U>(
-  fn: (value: T) => U
-) => (maybe: Maybe<T>): Maybe<U> => {
-  return maybe.map(fn);
-};
+export const map =
+  <T, U>(fn: (value: T) => U) =>
+  (maybe: Maybe<T>): Maybe<U> => {
+    return maybe.map(fn);
+  };
 
-export const chainM = <T, U>
-  (fn: (value: T) => Maybe<U>
-) => (maybe: Maybe<T>): Maybe<U> => {
-  return maybe.chain(fn);
-};
+export const chain =
+  <T, U>(fn: (value: T) => Maybe<U>) =>
+  (maybe: Maybe<T>): Maybe<U> => {
+    return maybe.chain(fn);
+  };
 
-export const unwrapOrM = <T>(
-  defaultValue: T
-) => (maybe: Maybe<T>): T => {
-  return maybe.unwrapOr(defaultValue);
-};
+export const unwrapOr =
+  <T>(defaultValue: T) =>
+  (maybe: Maybe<T>): T => {
+    return maybe.unwrapOr(defaultValue);
+  };
 
-export const matchM = <T, U>(
-  cases: { just: (value: T) => U; nothing: () => U }
-) => (maybe: Maybe<T>): U => {
-  return maybe.match(cases);
-};
+export const match =
+  <T, U>(cases: { just: (value: T) => U; nothing: () => U }) =>
+  (maybe: Maybe<T>): U => {
+    return maybe.match(cases);
+  };
 
-export const equalsM = <T>(
-  other: Maybe<T>
-) => (maybe: Maybe<T>): boolean => {
-  return maybe.equals(other);
-};
+export const equals =
+  <T>(other: Maybe<T>) =>
+  (maybe: Maybe<T>): boolean => {
+    return maybe.equals(other);
+  };
 
 export const just = <T>(value: T): Maybe<T> => new Just(value);
 export const nothing = <T = never>() => new Nothing<T>();